requirements.txt
biopython==1.81
pandas==2.1.3
py3Dmol==2.0.4
py3Dmol==2.0.4
stmol==0.0.9
streamlit==1.28.1
ipython-genutils==0.2.0
